+++ src/com/javaweb/dao/ParkingSpotPriceDAOImpl.java
@@ -9,7 +9,7 @@ import com.javaweb.po.ParkingSpotPrice;
 
 public class ParkingSpotPriceDAOImpl extends BaseDAOImpl<ParkingSpotPrice> implements ParkingSpotPriceDAO{
 	
-	public int findPricebySpotClassification(String classification){
+	public float findPricebySpotClassification(String classification){
 		List<ParkingSpotPrice> parkingSpotPrice=this.queryByAttribute(ParkingSpotPrice.class, "classification", classification);
 		
 		if(parkingSpotPrice!=null){
